pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "mugwort"
version = "0.3.1"
description = "一套由各种小脚本堆砌而成的工具集，主要用于爬虫和数据治理。"
readme = "README.md"
requires-python = ">=3.6.2,<4"
license = { file = "LICENSE" }
authors = [
    { name = "YongJie-Xie", email = "fsswxyj@qq.com" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Chinese (Simplified)",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Software Development",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "colorama",
]

[project.urls]
"Source" = "https://github.com/YongJie-Xie/MugwortTools"
"Tracker" = "https://github.com/YongJie-Xie/MugwortTools/issues"

[project.optional-dependencies]
all = [
    "cryptography",
    "requests",
    "pyyaml",
    "apscheduler",
]
crypto = [
    "cryptography",
]
proxy = [
    "requests",
    "pyyaml",
    "apscheduler",
]
test = [
    "pytest",
    "pytest-cov",
]
